requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
channels>=4.1
channels-redis>=4.0
psycopg2-binary>=2.9
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.2
celery>=5.4
redis>=4.5
python-dotenv>=1.0
gunicorn>=21.0
